working on setting mm_table

+++ arch/x86_64/include/asm/asm-string.h
@@ -1,7 +1,7 @@
 #ifndef _ASM_STRING_H_
 #define _ASM_STRING_H_
 
-static inline void *memcpy(void *pDest, const void *pSrc, size_t cb)
+static inline void *memcpy(void *pDest, const void *pSrc, uint64_t cb)
 {
   __asm__ volatile("cld; rep movsb"
                    : "=c" (cb), "=D" (pDest), "=S" (pSrc)
